archived_scripts/requirements.txt
anthropic>=0.49.0
apscheduler>=3.11.0
beautifulsoup4>=4.13.3
email-validator>=2.2.0
flask>=3.1.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.2
flask-dance
gunicorn>=23.0.0
mistralai>=1.6.0
numpy>=2.2.4
oauthlib
playwright>=1.51.0
psycopg2-binary>=2.9.10
psutil
pyjwt
requests
sqlalchemy>=2.0.40
trafilatura>=2.0.0
werkzeug>=3.1.3
wtforms>=3.2.1
matplotlib
numpy
scikit-learn
scipy
seaborn
matplotlib
numpy
scikit-learn
seaborn
trafilatura
openpyxl
pandas
trafilatura
playwright
anthropic
flask-login
flask-wtf
beautifulsoup4
imagehash
pillow
imagehash
